refactor(tests): Extract massive mock_dependencies ExitStack into shared conftest
